rename src folder, add favicon, update port
