+++ example.js
@@ -1,4 +1,4 @@
-import Crumbs from "./src/main.js";
+import Crumbs from "./lib/main.js";
 
 document.addEventListener("DOMContentLoaded", () => {
   const editCookieButton = document.querySelector(".edit-cookies");
