+++ index.html
@@ -4,7 +4,8 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./src/styles/main.css" />
+    <link rel="stylesheet" href="./lib/styles/main.css" />
+    <link rel="icon" href="/favicon.ico" type="image/x-icon" />
     <title>Crumbs</title>
   </head>
   <body>
@@ -24,7 +25,7 @@
     </button>
 
     <!-- <script src="./dist/crumbs.umd.js"></script> -->
-    <script type="module" src="./src/main.js"></script>
+    <script type="module" src="./lib/main.js"></script>
     <script type="module" src="example.js"></script>
   </body>
 </html>
