+++ lib/components/cookieBanner.js
