+++ lib/components/eventEmitter.js
