+++ lib/main.js
