+++ vite.config.js
@@ -4,9 +4,12 @@ import { defineConfig } from "vite";
 export default defineConfig({
   build: {
     lib: {
-      entry: path.resolve(__dirname, "src/main.js"),
+      entry: path.resolve(__dirname, "lib/main.js"),
       name: "Crumbs",
       fileName: "crumbs",
     },
+  },
+  server: {
+    port: 3000,
   },
 });
